requirements.txt
openai==2.7.2
pandas==2.3.3
python-dotenv==1.2.1
python-Levenshtein==0.27.3
tqdm==4.67.1
